+++ CHANGES.md
@@ -1,4 +1,8 @@
 # Changelog
 
 ## 0.1
+
 - Initial release
+
+- Add a feature to edit existing cards (Change the title and description)
+- Install @expo/vector-icons to add a edit icon in the Card.js